Move sort inputs to core
